@@ -99,7 +99,7 @@ abstract class CryptoBase
             }
             $pubKey = $pubKey->getKey(KeyPair::PUB_KEY);
         } elseif (!is_string($pubKey)) {
-            throw new InvalidArgumentException(sprintf('Encryption-public-key must be string of length %d bytes long, but is of type %s.', SODIUM_CRYPTO_BOX_PUBLICKEYBYTES, is_object($pubKey) ? get_class($pubKey) : gettype($pubKey)), 400);
+            throw new InvalidArgumentException(sprintf('Encryption-public-key must be string of length %d bytes long, but is of type %s.', SODIUM_CRYPTO_BOX_PUBLICKEYBYTES, get_debug_type($pubKey)), 400);
         } elseif (mb_strlen($pubKey, '8bit') !== SODIUM_CRYPTO_BOX_PUBLICKEYBYTES) {
             throw new InvalidArgumentException(sprintf('Encryption-public-key must be string of length %d bytes long, but is %d bytes.', SODIUM_CRYPTO_BOX_PUBLICKEYBYTES, mb_strlen($pubKey, '8bit')), 400);
         }
