@@ -1,12 +1,25 @@
-<?php declare(strict_types = 1);
+<?php declare(strict_types=1);
 
 namespace ricwein\Crypto\Tests\Asymmetric\File;
 
+use Exception;
 use PHPUnit\Framework\TestCase;
 use ricwein\Crypto\Asymmetric\KeyPair;
 use ricwein\Crypto\Asymmetric\Crypto;
+use ricwein\Crypto\Exceptions\EncodingException;
+use ricwein\Crypto\Exceptions\InvalidArgumentException;
+use ricwein\Crypto\Exceptions\MacMismatchException;
+use ricwein\Crypto\Exceptions\UnexpectedValueException;
+use ricwein\FileSystem\Exceptions\AccessDeniedException as FileSystemAccessDeniedException;
+use ricwein\FileSystem\Exceptions\ConstraintsException as FileSystemConstraintsException;
+use ricwein\FileSystem\Exceptions\Exception as FileSystemException;
+use ricwein\FileSystem\Exceptions\FileNotFoundException;
+use ricwein\FileSystem\Exceptions\RuntimeException as FileSystemRuntimeException;
+use ricwein\FileSystem\Exceptions\UnexpectedValueException as FileSystemUnexpectedValueException;
+use ricwein\FileSystem\Exceptions\UnsupportedException as FileSystemUnsupportedException;
 use ricwein\FileSystem\File;
 use ricwein\FileSystem\Storage;
+use SodiumException;
 
 /**
  * test symmetric message en/decryption
@@ -15,6 +28,10 @@ class DiskEncryptionTest extends TestCase
 {
     /**
      * @return File
+     * @throws FileSystemAccessDeniedException
+     * @throws FileSystemConstraintsException
+     * @throws FileSystemException
+     * @throws Exception
      */
     protected function getSourceFile(): File
     {
@@ -26,38 +43,62 @@ class DiskEncryptionTest extends TestCase
 
     /**
      * @return void
+     * @throws FileSystemAccessDeniedException
+     * @throws FileSystemConstraintsException
+     * @throws FileSystemException
+     * @throws SodiumException
+     * @throws EncodingException
+     * @throws InvalidArgumentException
+     * @throws MacMismatchException
+     * @throws UnexpectedValueException
+     * @throws FileNotFoundException
+     * @throws FileSystemRuntimeException
+     * @throws FileSystemUnexpectedValueException
+     * @throws FileSystemUnsupportedException
      */
-    public function testEncryption()
+    public function testEncryption(): void
     {
         $sourceFile = $this->getSourceFile();
         $destinationCipher = new Storage\Disk\Temp;
         $destinationPlain = new Storage\Disk\Temp;
 
-        $keypairA = (new KeyPair)->keygen();
-        $keypairB = (new KeyPair)->keygen();
+        $keypairA = KeyPair::generate();
+        $keypairB = KeyPair::generate();
 
         // encrypt
         $cipherFile = (new Crypto($keypairA))->encryptFile($sourceFile, $destinationCipher, $keypairB);
 
-        $this->assertSame($destinationCipher->path()->real, $cipherFile->path()->real);
-        $this->assertNotSame($sourceFile->read(), $cipherFile->read());
+        self::assertSame($destinationCipher->path()->real, $cipherFile->path()->real);
+        self::assertNotSame($sourceFile->read(), $cipherFile->read());
 
         // decrypt
         $plainFile = (new Crypto($keypairB))->decryptFile($cipherFile, $destinationPlain, $keypairA);
 
-        $this->assertSame($destinationPlain->path()->real, $plainFile->path()->real);
-        $this->assertSame($sourceFile->read(), $plainFile->read());
+        self::assertSame($destinationPlain->path()->real, $plainFile->path()->real);
+        self::assertSame($sourceFile->read(), $plainFile->read());
     }
 
     /**
      * @return void
+     * @throws EncodingException
+     * @throws FileNotFoundException
+     * @throws FileSystemAccessDeniedException
+     * @throws FileSystemConstraintsException
+     * @throws FileSystemException
+     * @throws FileSystemRuntimeException
+     * @throws FileSystemUnexpectedValueException
+     * @throws FileSystemUnsupportedException
+     * @throws InvalidArgumentException
+     * @throws MacMismatchException
+     * @throws SodiumException
+     * @throws UnexpectedValueException
      */
-    public function testSelfEncryption()
+    public function testSelfEncryption(): void
     {
         $sourceFile = $this->getSourceFile();
 
-        $keypairA = (new KeyPair)->keygen();
-        $keypairB = (new KeyPair)->keygen();
+        $keypairA = KeyPair::generate();
+        $keypairB = KeyPair::generate();
 
         $comparePath = $sourceFile->path()->real;
         $compareFile = $sourceFile->copyTo(new Storage\Disk\Temp);
@@ -65,13 +106,13 @@ class DiskEncryptionTest extends TestCase
         // encrypt
         $cipherFile = (new Crypto($keypairA))->encryptFile($sourceFile, null, $keypairB);
 
-        $this->assertSame($comparePath, $cipherFile->path()->real);
-        $this->assertNotSame($compareFile->read(), $cipherFile->read());
+        self::assertSame($comparePath, $cipherFile->path()->real);
+        self::assertNotSame($compareFile->read(), $cipherFile->read());
 
         // decrypt
         $plainFile = (new Crypto($keypairB))->decryptFile($cipherFile, null, $keypairA);
 
-        $this->assertSame($comparePath, $plainFile->path()->real);
-        $this->assertSame($plainFile->read(), $compareFile->read());
+        self::assertSame($comparePath, $plainFile->path()->real);
+        self::assertSame($plainFile->read(), $compareFile->read());
     }
 }
