@@ -12,7 +12,7 @@ use ricwein\Crypto\Symmetric\Crypto;
 use ricwein\FileSystem\Exceptions\AccessDeniedException as FileSystemAccessDeniedException;
 use ricwein\FileSystem\Exceptions\ConstraintsException as FileSystemConstraintsException;
 use ricwein\FileSystem\Exceptions\Exception as FileSystemException;
-use ricwein\FileSystem\Exceptions\FileNotFoundException;
+use ricwein\FileSystem\Exceptions\FileNotFoundException as FileSystemFileNotFoundException;
 use ricwein\FileSystem\Exceptions\RuntimeException as FileSystemRuntimeException;
 use ricwein\FileSystem\Exceptions\UnexpectedValueException as FileSystemUnexpectedValueException;
 use ricwein\FileSystem\Exceptions\UnsupportedException as FileSystemUnsupportedException;
@@ -23,7 +23,7 @@ use SodiumException;
 /**
  * test symmetric message en/decryption
  */
-class DiskEncryptionTest extends TestCase
+class MemoryEncryptionTest extends TestCase
 {
     /**
      * @return File
@@ -35,7 +35,7 @@ class DiskEncryptionTest extends TestCase
     protected function getSourceFile(): File
     {
         $message = base64_encode(random_bytes(random_int(2 << 9, 2 << 10)));
-        $file = new File(new Storage\Disk\Temp);
+        $file = new File(new Storage\Memory);
         $file->write($message);
         return $file;
     }
@@ -49,7 +49,7 @@ class DiskEncryptionTest extends TestCase
      * @throws InvalidArgumentException
      * @throws MacMismatchException
      * @throws UnexpectedValueException
-     * @throws FileNotFoundException
+     * @throws FileSystemFileNotFoundException
      * @throws FileSystemRuntimeException
      * @throws FileSystemUnexpectedValueException
      * @throws FileSystemUnsupportedException
@@ -57,29 +57,27 @@ class DiskEncryptionTest extends TestCase
     public function testEncryption(): void
     {
         $sourceFile = $this->getSourceFile();
-        $destinationCipher = new Storage\Disk\Temp;
-        $destinationPlain = new Storage\Disk\Temp;
+        $destinationCipher = new Storage\Memory;
+        $destinationPlain = new Storage\Memory;
         $key = Key::generate();
 
         // encrypt
         $cipherFile = (new Crypto($key))->encryptFile($sourceFile, $destinationCipher);
 
-        self::assertSame($destinationCipher->path()->real, $cipherFile->path()->real);
         self::assertNotSame($sourceFile->read(), $cipherFile->read());
 
         // decrypt
         $plainFile = (new Crypto($key))->decryptFile($cipherFile, $destinationPlain);
 
-        self::assertSame($destinationPlain->path()->real, $plainFile->path()->real);
         self::assertSame($sourceFile->read(), $plainFile->read());
     }
 
     /**
      * @return void
-     * @throws FileNotFoundException
      * @throws FileSystemAccessDeniedException
      * @throws FileSystemConstraintsException
      * @throws FileSystemException
+     * @throws FileSystemFileNotFoundException
      * @throws FileSystemRuntimeException
      * @throws FileSystemUnexpectedValueException
      * @throws FileSystemUnsupportedException
@@ -93,19 +91,16 @@ class DiskEncryptionTest extends TestCase
         $sourceFile = $this->getSourceFile();
         $key = Key::generate();
 
-        $comparePath = $sourceFile->path()->real;
-        $compareFile = $sourceFile->copyTo(new Storage\Disk\Temp);
+        $compareFile = $sourceFile->copyTo(new Storage\Memory);
 
         // encrypt
         $cipherFile = (new Crypto($key))->encryptFile($sourceFile);
 
-        self::assertSame($comparePath, $cipherFile->path()->real);
         self::assertNotSame($compareFile->read(), $cipherFile->read());
 
         // decrypt
         $plainFile = (new Crypto($key))->decryptFile($cipherFile);
 
-        self::assertSame($comparePath, $plainFile->path()->real);
         self::assertSame($plainFile->read(), $compareFile->read());
     }
 }
