@@ -1,12 +1,24 @@
-<?php declare(strict_types = 1);
+<?php declare(strict_types=1);
 
 namespace ricwein\Crypto\Tests\Symmetric\File;
 
+use Exception;
 use PHPUnit\Framework\TestCase;
+use ricwein\Crypto\Exceptions\InvalidArgumentException;
+use ricwein\Crypto\Exceptions\MacMismatchException;
+use ricwein\Crypto\Exceptions\UnexpectedValueException;
 use ricwein\Crypto\Symmetric\Key;
 use ricwein\Crypto\Symmetric\Crypto;
+use ricwein\FileSystem\Exceptions\AccessDeniedException as FileSystemAccessDeniedException;
+use ricwein\FileSystem\Exceptions\ConstraintsException as FileSystemConstraintsException;
+use ricwein\FileSystem\Exceptions\Exception as FileSystemException;
+use ricwein\FileSystem\Exceptions\FileNotFoundException as FileSystemFileNotFoundException;
+use ricwein\FileSystem\Exceptions\RuntimeException as FileSystemRuntimeException;
+use ricwein\FileSystem\Exceptions\UnexpectedValueException as FileSystemUnexpectedValueException;
+use ricwein\FileSystem\Exceptions\UnsupportedException as FileSystemUnsupportedException;
 use ricwein\FileSystem\File;
 use ricwein\FileSystem\Storage;
+use SodiumException;
 
 /**
  * test symmetric message en/decryption
@@ -15,6 +27,7 @@ class MixedEncryptionTest extends TestCase
 {
     /**
      * @return string
+     * @throws Exception
      */
     protected function getMessage(): string
     {
@@ -23,42 +36,66 @@ class MixedEncryptionTest extends TestCase
 
     /**
      * @return void
+     * @throws SodiumException
+     * @throws InvalidArgumentException
+     * @throws MacMismatchException
+     * @throws UnexpectedValueException
+     * @throws FileSystemAccessDeniedException
+     * @throws FileSystemConstraintsException
+     * @throws FileSystemException
+     * @throws FileSystemFileNotFoundException
+     * @throws FileSystemRuntimeException
+     * @throws FileSystemUnexpectedValueException
+     * @throws FileSystemUnsupportedException
+     * @throws Exception
      */
-    public function testMemoryToDisk()
+    public function testMemoryToDisk(): void
     {
-        $key = (new Key)->keygen();
+        $key = Key::generate();
         $message = $this->getMessage();
 
         $source = (new File(new Storage\Memory))->write($message);
         $cipher = (new Crypto($key))->encryptFile($source, new Storage\Disk\Temp);
 
-        $this->assertInstanceOf(Storage\Disk\Temp::class, $cipher->storage());
+        self::assertInstanceOf(Storage\Disk\Temp::class, $cipher->storage());
 
         $plainA = (new Crypto($key))->decryptFile($cipher, new Storage\Memory);
         $plainB = (new Crypto($key))->decryptFile($cipher, new Storage\Disk\Temp);
 
-        $this->assertSame($message, $plainA->read());
-        $this->assertSame($message, $plainB->read());
+        self::assertSame($message, $plainA->read());
+        self::assertSame($message, $plainB->read());
     }
 
 
     /**
      * @return void
+     * @throws FileSystemAccessDeniedException
+     * @throws FileSystemConstraintsException
+     * @throws FileSystemException
+     * @throws FileSystemFileNotFoundException
+     * @throws FileSystemRuntimeException
+     * @throws FileSystemUnexpectedValueException
+     * @throws FileSystemUnsupportedException
+     * @throws InvalidArgumentException
+     * @throws MacMismatchException
+     * @throws SodiumException
+     * @throws UnexpectedValueException
+     * @throws Exception
      */
-    public function testDiskToMemory()
+    public function testDiskToMemory(): void
     {
-        $key = (new Key)->keygen();
+        $key = Key::generate();
         $message = $this->getMessage();
 
         $source = (new File(new Storage\Disk\Temp))->write($message);
         $cipher = (new Crypto($key))->encryptFile($source, new Storage\Memory);
 
-        $this->assertInstanceOf(Storage\Memory::class, $cipher->storage());
+        self::assertInstanceOf(Storage\Memory::class, $cipher->storage());
 
         $plainA = (new Crypto($key))->decryptFile($cipher, new Storage\Memory);
         $plainB = (new Crypto($key))->decryptFile($cipher, new Storage\Disk\Temp);
 
-        $this->assertSame($message, $plainA->read());
-        $this->assertSame($message, $plainB->read());
+        self::assertSame($message, $plainA->read());
+        self::assertSame($message, $plainB->read());
     }
 }
